Move components to its proper folder
